QuickJS: disable workers
Workers are not suitable for testing with libFuzzer due to their
multi-file nature: each libFuzzer corpus item is independent and
processed in isolation. As a result, it's not feasible to ensure that
any referenced worker file is available or remains accessible during
a fuzzing session.

Additionally, if the fuzz target references a nonexistent worker file
and any handler is attached to it, it can cause the process to hang
indefinitely.

This patch disables worker support during fuzzing to avoid such issues.