ffmpeg: adjust build config of deps to avoid upstream bugs
freetype added dynamic loading of harfbuzz, but it doesn't built, either
way we won't use it, so just disable harfbuzz for freetype. And if you
wander why we don't install harfbuzz first, it's because of cyclic
dependency, harfbuzz depends on freetype and would built it's own copy
of it as a subproject.

While at it disable tests and tools in fontconfig build, not needed.