index_build: Build indexes for non-fuzzing targets, if needed
This adds the flag `--target-type` to the indexer script, with
possible values "fuzzing" and "all".  When set to "fuzzing", the clang
wrapper script and the indexer will continue to only index fuzzing
targets.  When set to "all", all binary targets will be indexed.

This can be used together with the `--targets` flag to index specific
command line binaries which are not fuzzing targets.

The flag is passed down to the clang wrapper script in the form of the
`INDEXER_TARGET_TYPE` environment variable.