[infra] Indicate coverage run failure when libFuzzer reports an error (#13103)
Currently OOM's and timeouts for individual corpus entries during
coverage measurement are silently ignored, as libFuzzer still exits with
a zero returncode. This PR adds a grep for error messages during the
coverage run. Additionally, add a small log file is created to indicate
the fuzzer target for which the coverage measurement did not complete
succesfully.

---------

Signed-off-by: phi-go <dev@philipp-goerz.com>
Co-authored-by: DavidKorczynski <david@adalogics.com>