index_build: Add `--targets-all-index`.
When used together with `--targets`, this loosens the filtering a bit to still
allow all binaries to be indexed.

We need this because some projects use different binary names during build time
compared to what's copied to $OUT/.

The `--targets` list is only taken into account when we are producing archives,
which still saves a fair bit of time for projects with lots of them.

PiperOrigin-RevId: 783542238