imagemagick: add replay_build.sh (#13692)
Adds a targeted version of `replay_build.sh`. Targeted meaning that the
first argument to the bash script (`$1`) is the only harness that is
recompiled.

We rely on the `make_build_replayable.py` wrappers for all of the
recompilation except the harness itself. Only a single harness is build.

```
imagemagick: Compile times: Vanilla=356; Replay=47;
```

Signed-off-by: David Korczynski <david@adalogics.com>