Update jdk 17 version to unbreak the PDFBox build (#13860)
This updates the version of jdk17 to 17.0.16+8. 

This update is required to build PDFBox:
https://issues.apache.org/jira/browse/PDFBOX-6051?focusedCommentId=18015721&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-18015721

PDFBox is no longer building in ossfuzz: 

https://oss-fuzz-build-logs.storage.googleapis.com/log-907a41b7-64fc-4a03-a728-a4b5bbf8042a.txt

and the error message is consistent with PDFBOX-6051:

```
Fatal error compiling: error: invalid flag: -proc:full
```

I confirmed locally that `infra/base-images/all.sh` creates a jvm base
that in turn allows pdfbox to build.

I have NOT tested this across all projects.

---------

Co-authored-by: DavidKorczynski <david@adalogics.com>