Fix(infra): Set GCB build context to base-images dir
The previous change caused the GCB builds to fail because the build
context was set to the root of the OSS-Fuzz checkout. This failed
because the Dockerfiles in infra/base-images rely on files relative
to that directory.

This change fixes this by explicitly setting the build context in the
'gcloud builds submit' command to the 'infra/base-images' directory.