[elfutils] no longer send notifications to the elfutils mailing list (#13997)
as requested in
https://sourceware.org/pipermail/elfutils-devel/2025q3/008514.html

I can keep track of bugs the fuzz targets can discover elsewhere so
there is no need to keep my email address there either.

Ideally it would be great if bugs were reported to the elfutils bug
tracker. It would also be useful if backtraces were attached to bug
reports on https://issues.oss-fuzz.com/ so that it wouldn't be necessary
to have Google accounts to view them on https://oss-fuzz.com.
view_restrictions is none there so all the stuff related to disclosures
isn't relevant.