Install Java 25 in runners The latest Java version available on oss-fuzz is 17. Some Java frameworks and libraries are now moving beyond Java 17 as the earliest supported version, mostly to Java 21.
This PR adds Java 25 to the base-runner and base-builder-jvm containers, under JAVA_25_HOME. This should be a reasonable option for a few years.

I've chosen corretto as the OpenJDK build because temurin is apparently not yet available for normal linux, and because corretto has a permalink that always downloads the latest Java 25 version, so we will get bugfixes without changing the install scripts further.