go coverage: align covermode and run only coverage test (#14052)
This PR does so that the go coverage build only runs the coverage test
in the compiled binary. It does so by adding the `-test.run` flag when
running go-fuzz fuzzers and by also storing the function name of go-fuzz
fuzzers in `$OUT/fuzzer_function_names.json`. Also, it limits the
function name in `-test.run` to be specific to the actual name, since
this is a regex and can include other functions depending on how they
are named.

This PR also sets the go-fuzz fuzzers' covermode to `atomic` so that all
coverage binaries are atomic. This avoids this failure:
https://github.com/wadey/gocovmerge/blob/b5bfa59ec0adc420475f97f89b58045c721d761c/gocovmerge.go#L18.

This PR also removes `-run Test${function}Corpus` from
`compile_go_fuzzer` since it doesn't do anything. Adding the flag when
compiling the binary will not limit the compilation to only the
specified test.

Signed-off-by: Adam Korczynski <adam@adalogics.com>