nodejs: migrate fuzzers to oss-fuzz (#14143)
Since last year, I have maintained many of Node.js's fuzzers from a fork
of mine to ensure they ran over a longer period of time. This PR
migrates the fuzzers I've had on my fork over to OSS-Fuzz. Their final
destination should be upstream, but we may still need to iterate over
these a few times which is why OSS-Fuzz is a good place for now.

The PR includes the fuzzers, fuzz helpers and a script that adds the
fuzzers to node.js's `node.gyp` file so they can be built in node.js's
build system.

The PR also removes the sha from the base-builder in the Dockerfile
since it breaks the build.

---------

Signed-off-by: Adam Korczynski <adam@adalogics.com>