Copy missing fuzzers to image (#14253)
Depends on https://github.com/hugovk/python-library-fuzzers/pull/1 to
avoid errors.
This pull request copies the following missing fuzzers into the image:

* `fuzzer-re`
* `fuzzer-zipfile`
* `fuzzer-tarfile`
* `fuzzer-configparser`
* `fuzzer-tomllib`
* `fuzzer-plistlib`

Need to create a mechanism that ensures this doesn't happen in the
future. Also some documentation, as it's not obvious when creating a new
fuzzer on the `python-library-fuzzers` repository that you'd need to
update this `build.sh` file, too.