fix: Unbreak `neqo` compilation (#14319)
As described in https://github.com/mozilla/neqo/issues/3147, we made
changes to assume a release build of NSS to be present for neqo, which
is failing in `oss-fuzz` due to toolchain issues. Work around this.

Also switch to pulling NSPR from the Mercurial repository, as the GitHub
mirror is not being updated it seems.

Finally, build NSS like we do for neqo, modulo not doing a release
build.