fix(infra): Revert trial build logic to select successful projects (#14440)
This PR reverts the logic change introduced in
cabb69336c0c5eab14099d2f078000daafa5fc91. The intention of trial builds
is to verify that changes to base images do not break existing healthy
projects. The regression caused the script to select failed projects
instead, which defeats the purpose of regression testing.