airflow: fix build (#14326)
The reason for the airflow fuzzing build error is that pip attempts to
download airflow-core from PyPI, but cannot find the locally developed
version (3.2.0+), thus resulting in the error. The new solution provides
the appropriate version of airflow-core to solve the problem by
installing./task-sdk,./airflow core and the current directory.