PDFBox add harnesses and seeds (#13873)
This adds harnesses for ExtractText and font parsers. It also adds seeds
from an arbitrary zip of 1k pdfs from
https://digitalcorpora.org/corpora/file-corpora/cc-main-2021-31-pdf-untruncated/.

Along the way, this also updates maven, and it makes a small
modification to allow local builds (which did not work before). Finally,
this adds log4j2 as the logging implementation and turns logging off to
avoid corrupting the console.

I went with `fuzzerTestOneInput()` rather than the `@FuzzTest`
annotation because I couldn't get `reproduce` to work with the
`@FuzzTest` annotation. This could be an unrelated issue or user error.